code/backend/requirements.txt
# Enterprise ITSM Graph RAG Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
astrapy==1.5.2
ibm-watsonx-ai==1.1.15
httpx==0.27.2
